Panic if we got too many failed decodes
I'm getting occasional issues when the app just fails to decode the
signal for quite some time. If controller is restarted it works fine
again. Panic if we get too many failed decodes as a workaround for now
until actual root cause is found (or rather if it ever found)