Fix handling negative zero temp and validate temp range
Temperature sign needs to be extracted explicitly, otherwise it will be
lost when the integer part is extracted using division by ten, since integer
zero has no sign.

Also validate temperature range. Nexus-TH doesn't have any checksum to
validate the payload, so the values that are not plausible need to be
rejected